feat: define createRandomPeers() & change the name of file displayPeers.ts
